REFACTOR: Move squares to new `CommutativeDiagrams` module.
